Add "tmp/console" to .gitignore
The equiavlent to running a console in Lucky requires using `lucky exec`
command which creates a file in temp/console for you to edit and
execute. Since Crystal is a compiled language, it cannot execute live
through an interpreter. Sometimes this code contains sensitive info
as various scenarios are tested. Removing it from git to ensure
current and future secrets are not accidentally passed into the
repository.